Create <AddPet /> component
