+++ web/src/components/PetList/components/AddPet/index.js
@@ -1,0 +1,1 @@
+export * from './AddPet';
